Refactored `Logger` trait to support asynchronous logging
- The `Logger` trait has been refactored to support asynchronous logging, enhancing the safe use of the `log_adapter` between frontend and backend.
- An `async` keyword has been added to the `log` function within the trait.
- A `#[tokio::main]` annotation has been put to the main function as this function now involves asynchronous logging.
- This refactor is crucial in facilitating non-blocking operations, thereby improving the overall application's performance.
