+++ fullstack_rust_app/frontend/src/lib.rs
@@ -1,0 +1,8 @@
+pub mod adapters {
+    /// Module providing the WebServerAdapter functionality.
+    ///
+    /// This module contains the code related to the adapter for the web server.
+    /// It provides an interface for starting and stopping the web server, as well
+    /// as handling incoming HTTP requests and responses.
+    pub mod web_server_adapter;
+}
